added new files

+++ src/Counter.sol
@@ -4,7 +4,7 @@ pragma solidity 0.8.17;
 contract Counter {
     uint256 public count;
 
-    function getCount() public view returns (uint) {
+    function getCount() public view returns (uint256) {
         return count;
     }
